Fixed bug: Normal vector average and colours
The colourization of normal vectors has been changed to map [-1, 1] to
[0, 1].
Also, there was a bug that was causing the colour of the normal vector
to be incorrect. The reason was that in the function computeNormalsNeil,
the normal vectors were always computed as the average of six elements
which it is not always true. Now, the normal will be computed as the sum
of all the face normal vectors and then normalized.