ply IO
Functionality to read and write ply files added (using original
implementation from Greg Turk, March 1994 -
http://www.dcs.ed.ac.uk/teaching/cs4/www/graphics/Web/ply.html)
Reads plain text and binary ply files. These are required to have float
xyz vertex coordinates and unsigned char rgb color. It supports also
rgba and vertex normals (THIS LAST ONE IS PARSED BUT NOT CURRENTLY
LOADED).
When writing, it generates a ply file with float xyz coordinates, float
vertex normal vectors and unsigned char vertex rgb color. IThis is saved
as a binary file since there is not interest in save it as plain text.
If this want to be changed in the futute, the funcionality for it is
already implemented. just need to change the part of the code where
checks the definition of PLY_SAVE_ASCII.